extract deployment-edit-modal to new component deployments-edit-modal
